requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
distro==1.9.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
install==1.3.5
numpy==1.26.3
openai==1.7.2
openi==1.2.2
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
regex==2023.12.25
requests==2.31.0
scipy==1.11.4
sniffio==1.3.0
tenacity==8.2.3
termcolor==2.4.0
tiktoken==0.5.2
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
gunicorn
flask
requests_oauthlib
flask_cors
